requirements.txt
cloudpickle==3.0.0
contourpy==1.2.0
cycler==0.12.1
distlib==0.3.8
filelock==3.13.4
fonttools==4.47.0
gym==0.26.2
gym-notices==0.0.8
joblib==1.3.2
kiwisolver==1.4.5
matplotlib==3.8.2
networkx==3.2.1
numpy==1.26.2
packaging==23.2
pbr==6.0.0
Pillow==10.1.0
platformdirs==4.2.0
pydot==1.4.2
pygame==2.6.0
pyparsing==3.1.1
python-dateutil==2.8.2
scikit-learn==1.3.2
scipy==1.11.4
six==1.16.0
stevedore==5.2.0
threadpoolctl==3.2.0
virtualenv==20.25.1
virtualenv-clone==0.5.7
virtualenvwrapper==6.1.0
